requirements.txt
Flask==2.3.6
requests==2.31.0
gunicorn==21.2.0
google-genai==1.50.1

